refactor(zim): finalize zim_operations.py as backward-compat shim
Move ZimOperations, zim_archive, PaginationCursor, the constants, and
the remaining coordinator methods (file listing, metadata extraction,
main-page lookup, _resolve_entry_with_fallback) from
zim_operations.py into the new openzim_mcp.zim.archive module. The
zim_operations module now does nothing but re-export the public and
test-patched symbols so existing imports and patches keep working.

Also wire up zim/__init__.py to re-export from archive.py.

The two methods on ZimOperations that open archives directly
(get_zim_metadata, get_main_page) and the zim_archive context manager
itself look up Archive / zim_archive on the openzim_mcp.zim_operations
shim at call time, so test patches against those names propagate
through to the real call sites without test changes.
